+++ Models/LearnerQuestionMark.cs
@@ -10,6 +10,6 @@ namespace ARIS1.Models
         public AssessmentQuestion Question { get; set; } = null!;
         public int LearnerId { get; set; }
         public Learner Learner { get; set; } = null!;
-        public float MarksAwarded { get; set; }
+        public decimal MarksAwarded { get; set; }
     }
 }
